+++ R/download_bg_acs_raw.R
@@ -139,6 +139,7 @@ download_bg_acs_raw <- function(yr,
         format = stage_format,
         overwrite = overwrite,
         validation_strict = validation_strict,
+        yr = yr,
         storage = storage
       )
     }